+++ frontend_files/ci_project/ci_project/views.py
@@ -1,6 +1,8 @@
 from django.shortcuts import render
 from django.core.mail import send_mail
+from django.utils.html import strip_tags
 from ci_shop.models import Product
+from django.template.loader import render_to_string
 
 def home(request):
     data=Product.objects.filter(featured =True).order_by('-id')
@@ -14,14 +16,17 @@ def contactUs(request):
         subject = request.POST.get('input-subject')
         message = request.POST.get('message')
 
-        data = {
+        context = {
             'name': name,
             'email': email,
             'subject': subject,
-            'message': message
+            'message': message,
         }
-        message = '''\nNew message: {}\nFrom: {} {}'''.format(data['message'], data['name'], data['email'])
-        send_mail(data['subject'], message, "bounces+28347242@em9511.ctrlintel.shop", ['controlintel23@gmail.com '])
+        
+        html_message = render_to_string('contact_msg.html', context)
+        plain_text = strip_tags(html_message)
+
+        send_mail(context['subject'], plain_text, "bounces+28347242@em9511.ctrlintel.shop", ['controlintel23@gmail.com'], html_message=html_message)
     return render(request,'contact.html')
