+++ frontend_files/ci_project/templates/contact_msg.html
@@ -1,0 +1,62 @@
+<!DOCTYPE html>
+<html lang="en" xmlns:o="urn:schemas-microsoft-com:office:office" xmlns:v="urn:schemas-microsoft-com:vml">
+
+<head>
+    <title></title>
+    <meta content="text/html; charset=utf-8" http-equiv="Content-Type" />
+    <meta content="width=device-width,initial-scale=1" name="viewport" />
+    <!--[if mso]><xml><o:OfficeDocumentSettings><o:PixelsPerInch>96</o:PixelsPerInch><o:AllowPNG/></o:OfficeDocumentSettings></xml><![endif]-->
+    <!--[if !mso]><!-->
+    <link href="https://fonts.googleapis.com/css?family=Abril+Fatface" rel="stylesheet" type="text/css" />
+    <!--<![endif]-->
+</head>
+
+<body style="background-color:#fff;margin:0;padding:0;-webkit-text-size-adjust:none;text-size-adjust:none">
+    <table width="100%" style="width:100%;">
+        <tbody>
+            <tr>
+                <td>
+                    <table style="background-color:#fff8f5; width: 60%; margin:50px auto 0 auto;">
+                        <tbody>
+                            <tr>
+                                <td style="padding-bottom: 5px;">
+                                    <h1 style="text-align:center;">Ctrl  
+                                    <span style="color:orange">
+                                        Intelligence
+                                    </span>
+                                    </h1>
+                                    
+                                </td>
+                            </tr>
+                        </tbody>
+                    </table>
+                    
+
+                    <table style="background-color:#fff8f5; width: 60%; margin:0 auto 0 auto; padding: 0 0 10px 50px;">
+                        <tbody>
+                            <tr>
+                                <td>
+                                    <h2 style="text-align:center;">New Message from: {{name}}</h2>
+                                    <h2 style="text-align:center;">Email: {{email}}</h2>
+                                        <hr style="border-width: 4px; border-style: solid none none none; width:13%; color:#de7f50;">
+                                </td>
+                            </tr>
+                            <tr>
+                                <td>
+                                    <p>
+                                        {{message}}
+                                    </p>
+                                </td>
+                            </tr>
+                        </tbody>
+                    </table>
+                </td>
+            </tr>
+        </tbody>
+    </table>
+
+
+
+</body>
+
+</html>
